+++ udata/core/dataset/forms.py
@@ -12,6 +12,7 @@ from .models import (
     Dataset, Resource, License, Checksum, CommunityResource,
     UPDATE_FREQUENCIES, DEFAULT_FREQUENCY, RESOURCE_FILETYPES, CHECKSUM_TYPES,
     LEGACY_FREQUENCIES, RESOURCE_TYPES, RESOURCE_FILETYPE_FILE,
+    ResourceSchema,
 )
 
 __all__ = ('DatasetForm', 'ResourceForm', 'CommunityResourceForm')
@@ -48,6 +49,17 @@ def enforce_filetype_file(form, field):
         ))
 
 
+def enforce_allowed_schemas(form, field):
+    schema = field.data
+    allowed_schemas = [s['id'] for s in ResourceSchema.objects()]
+    if schema not in allowed_schemas:
+        message = _('Schema "{schema}" is not an allowed value. Allowed values: {values}')
+        raise validators.ValidationError(message.format(
+            schema=schema,
+            values=', '.join(allowed_schemas)
+        ))
+
+
 class BaseResourceForm(ModelForm):
     title = fields.StringField(_('Title'), [validators.DataRequired()])
     description = fields.MarkdownField(_('Description'))
@@ -79,6 +91,11 @@ class BaseResourceForm(ModelForm):
         _('Publication date'),
         description=_('The publication date of the resource'))
     extras = fields.ExtrasField()
+    schema = fields.StringField(
+        _('Schema'),
+        default=None,
+        validators=[validators.optional(), enforce_allowed_schemas],
+        description=_('The schema slug the resource adheres to'))
 
 
 class ResourceForm(BaseResourceForm):