+++ udata/tests/dataset/test_dataset_model.py
@@ -1,12 +1,19 @@
 from datetime import datetime, timedelta
 
 import pytest
+import requests
 
 from mongoengine import post_save
 
-from udata.models import db, Dataset, License, LEGACY_FREQUENCIES
+from udata.app import cache
+from udata.models import (
+    db, Dataset, License, LEGACY_FREQUENCIES, ResourceSchema
+)
 from udata.core.dataset.factories import (
     ResourceFactory, DatasetFactory, CommunityResourceFactory, LicenseFactory
+)
+from udata.core.dataset.exceptions import (
+    SchemasCatalogNotFoundException, SchemasCacheUnavailableException
 )
 from udata.core.discussions.factories import (
     MessageDiscussionFactory, DiscussionFactory
@@ -413,3 +420,44 @@ class LicenseModelTest:
         found = License.guess('should not match', license.id)
         assert isinstance(found, License)
         assert license.id == found.id
+
+
+class ResourceSchemaTest:
+    @pytest.mark.options(SCHEMA_CATALOG_URL='https://example.com/notfound')
+    def test_resource_schema_objects_404_endpoint(self):
+        with pytest.raises(SchemasCatalogNotFoundException):
+            ResourceSchema.objects()
+
+    @pytest.mark.options(SCHEMA_CATALOG_URL='https://example.com/schemas')
+    def test_resource_schema_objects_timeout_no_cache(self, client, rmock):
+        rmock.get('https://example.com/schemas', exc=requests.exceptions.ConnectTimeout)
+        with pytest.raises(SchemasCacheUnavailableException):
+            ResourceSchema.objects()
+
+    @pytest.mark.options(SCHEMA_CATALOG_URL='https://example.com/schemas')
+    def test_resource_schema_objects(self, app, rmock):
+        rmock.get('https://example.com/schemas', json={
+            'schemas': [{"name": "etalab/schema-irve", "title": "Schéma IRVE"}]
+        })
+
+        assert ResourceSchema.objects() == [{"id": "etalab/schema-irve", "label": "Schéma IRVE"}]
+
+    @pytest.mark.options(SCHEMA_CATALOG_URL=None)
+    def test_resource_schema_objects_no_catalog_url(self):
+        assert ResourceSchema.objects() == []
+
+    @pytest.mark.options(SCHEMA_CATALOG_URL='https://example.com/schemas')
+    def test_resource_schema_objects_w_cache(self, rmock, mocker):
+        cache_mock_set = mocker.patch.object(cache, 'set')
+        mocker.patch.object(cache, 'get', return_value='dummy_from_cache')
+
+        # fill cache
+        rmock.get('https://example.com/schemas', json={
+            'schemas': [{"name": "etalab/schema-irve", "title": "Schéma IRVE"}]
+        })
+        ResourceSchema.objects()
+        assert cache_mock_set.called
+
+        rmock.get('https://example.com/schemas', status_code=500)
+        assert 'dummy_from_cache' == ResourceSchema.objects()
+        assert rmock.call_count == 2